+++ src/localstorage/index.js
@@ -1,5 +1,5 @@
 if (window.angular && window.angular.env && window.angular.env.QT_UI_LOG) {
-  window.console.log('qt-angular-ui/src/wechat load');
+  window.console.log('qt-angular-ui/src/localstorage load');
 }
 
 import component   from './_index';